Move controller helpers to shared test-e2e directory (for now)
