+++ extensions/amp-base-carousel/test-e2e/helpers-amp.js
